Add files via upload
app.py: This script contains the Streamlit application for the frontend. Users can upload an image, and the app processes it to find the closest matching Bollywood celebrity.

feature_extractor.py: This script is responsible for extracting ResNet50 features from images. It processes the images stored in filenames.pkl and generates embeddings for each.

test.py: This script takes an uploaded image, extracts its features using ResNet50, and compares its cosine similarity to all the features previously extracted from the images in the database. It identifies the closest match based on cosine distances